Update README.md
The objective of a database course is to teach students the fundamental principles of information management, including the design, creation, and manipulation of databases. The goal is to provide a thorough understanding of concepts such as relational models, SQL queries, and normalization, enabling effective and secure data management in various professional contexts. Additionally, students learn practical skills to design database management systems tailored to the specific needs of organizations.